refactored upsert methods
