+++ schematic_db/rdb/postgres.py
@@ -1,9 +1,13 @@
 """Represents a Postgres database."""
 from typing import Any
+import numpy as np
 import sqlalchemy as sa
 import sqlalchemy.dialects.postgresql as sa_postgres
 import pandas as pd
+from sqlalchemy.inspection import inspect
+from sqlalchemy import exc
 from .sql_alchemy_database import SQLAlchemyDatabase, SQLConfig
+from .rdb import UpsertDatabaseError
 
 
 class PostgresDatabase(SQLAlchemyDatabase):
@@ -26,8 +30,32 @@ class PostgresDatabase(SQLAlchemyDatabase):
         """
         super().__init__(config, verbose, "postgresql")
 
+    def upsert_table_rows(self, table_name: str, data: pd.DataFrame) -> None:
+        """Inserts and/or updates the rows of the table
+
+        Args:
+            table_name (str): The name of the table to be upserted
+            data (pd.DataFrame): The rows to be upserted
+
+        Raises:
+            UpsertDatabaseError: Raised when a SQLAlchemy error caught
+        """
+        table = sa.Table(table_name, self.metadata, autoload_with=self.engine)
+        data = data.replace({np.nan: None})
+        rows = data.to_dict("records")
+        table_schema = self.metadata.tables[table_name]
+        primary_key = inspect(table_schema).primary_key.columns.values()[0].name
+        try:
+            self._upsert_table_rows(rows, table, table_name, primary_key)
+        except exc.SQLAlchemyError as exception:
+            raise UpsertDatabaseError(table_name) from exception
+
     def _upsert_table_rows(
-        self, rows: list[dict[str, Any]], table: sa.table, table_name: str, primary_key: str
+        self,
+        rows: list[dict[str, Any]],
+        table: sa.table,
+        table_name: str,
+        primary_key: str,
     ) -> None:
         """Upserts a pandas dataframe into a Postgres table
 
@@ -38,7 +66,9 @@ class PostgresDatabase(SQLAlchemyDatabase):
             primary_key (str): The name fo the primary key of the table being upserted into
         """
         statement = sa_postgres.insert(table).values(rows)
-        update_columns = {col.name: col for col in statement.excluded if col.name != primary_key}
+        update_columns = {
+            col.name: col for col in statement.excluded if col.name != primary_key
+        }
         statement = statement.on_conflict_do_update(
             constraint=f"{table_name}_pkey", set_=update_columns
         )
