+++ tests/test_sql.py
@@ -31,8 +31,7 @@ def fixture_sql_databases(
     postgres_database: PostgresDatabase,
 ) -> Generator:
     """Yields a list of databases to test"""
-    #yield [mysql_database, postgres_database]
-    yield [postgres_database]
+    yield [mysql_database, postgres_database]
 
 
 @pytest.mark.fast
